CDPR Dataset for Trajectory Tracing
The data set for the underactuated cable-driven parallel robot (CDPR) was used to calculate the natural oscillation frequencies of the end effector (EE) for the control purpose. The author conducted an experiment for an underactuated CDPR system with 2, 3 and 4-cable and 6-DOF to study the kinematics, dynamics, stability and stiffness of the system with the aim of implementing trajectory planning for a CDPR with 6-DOF, 4-cable using the frequency-based method. I found this data set useful for this work as I aim to design the motion control and trajectory planning to achieve stability, precision, and to maintain positive cable tensions. The author used the frequency-based method for their study, while I plan to explore the applications of reinforcement learning (RL) algorithms such as proximity policy optimization (PPO) to train and allow the model to learn from the dataset for trajectory tracing and new trajectory generation with the aim of achieving high precision, stability, and maintaining positive tensions throughout the multiple cables. The PPO algorithm is applied to a simplified CDPR system with four cables (4-cable) and six degrees of freedom (6-DOF) for motion control and trajectory planning. The model-free error compensations proposed by has further demonstrated the accuracy of the deep learning method for training dataset. The dataset for this work is obtained as open access through the IEEE Data Port.